api/requirements.txt
attrs==21.2.0
azure-functions==1.7.2
cffi==1.14.6
colorama==0.4.4
cryptography==3.4.7
dateutils==0.6.12
Deprecated==1.2.12
docopt==0.6.2
iniconfig==1.1.1
jwcrypto==1.0
mongoengine==0.23.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.4
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jwt==3.3.0
pytz==2021.1
six==1.16.0
toml==0.10.2
watchdog==2.1.3
wrapt==1.12.1
